test(FilesController): add files controller unit test file with json files recovery test method
