+++ tests/Ml.Cli.WebApp.Tests/ComparesControllerUnitTest.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using System.Threading.Tasks;
 using Ml.Cli.FileLoader;
 using Ml.Cli.WebApp.Controllers;
 using Ml.Cli.WebApp.Paths;
@@ -33,9 +31,7 @@ namespace Ml.Cli.WebApp.Tests
             var basePath = new Mock<BasePath>("");
             basePath.Setup(mock => mock.IsPathSecure(It.IsAny<string>())).Returns(true);
             
-            var comparesPath = new ComparesPaths("");
-
-            var compareController = new ComparesController(fileLoader.Object, basePath.Object, comparesPath);
+            var compareController = new ComparesController(fileLoader.Object, basePath.Object);
             await compareController.SaveCompare(editorContent);
             
             var compareResultContent = "{\"CompareLocation\":\"someFolder\\\\compareFile.json\",\"Content\":[{\"FileName\":\"{other_file}_pdf.json\",\"Left\":{\"Url\":\"https://url\",\"FileName\":\"{other_file}.pdf\",\"FileDirectory\":\"someFolder1\\\\{other_file}_pdf.json\",\"ImageDirectory\":null,\"FrontDefaultStringsMatcher\":null,\"StatusCode\":200,\"Body\":\"{\\\"analysis\\\":\\\"content_body_left_other_file\\\"}\",\"Headers\":[],\"TimeMs\":0,\"TicksAt\":0},\"Right\":{\"Url\":\"https://url\",\"FileName\":\"{other_file}.pdf\",\"FileDirectory\":\"someFolder2\\\\{other_file}_pdf.json\",\"ImageDirectory\":\"someFolderImages\",\"FrontDefaultStringsMatcher\":\"rotation\",\"StatusCode\":200,\"Body\":\"{\\\"analysis\\\":\\\"content_body_right_other_file\\\"}\",\"Headers\":[],\"TimeMs\":1,\"TicksAt\":1}},{\"FileName\":\"{fileName}_pdf.json\",\"Left\":{\"Url\":\"https://url\",\"FileName\":\"{fileName}.pdf\",\"FileDirectory\":\"someFolder1\\\\{fileName}_pdf.json\",\"ImageDirectory\":null,\"FrontDefaultStringsMatcher\":null,\"StatusCode\":200,\"Body\":\"{\\\"analysis\\\":\\\"content_body_left_fileName\\\"}\",\"Headers\":[],\"TimeMs\":0,\"TicksAt\":0},\"Right\":{\"Url\":\"https://url\",\"FileName\":\"{fileName}.pdf\",\"FileDirectory\":\"someFolder2\\\\{fileName}_pdf.json\",\"ImageDirectory\":\"someFolderImages\",\"FrontDefaultStringsMatcher\":\"rotation\",\"StatusCode\":200,\"Body\":\"{\\\"analysis\\\":\\\"content_body_httpResult\\\"}\",\"Headers\":[],\"TimeMs\":1,\"TicksAt\":1}}]}";
@@ -44,26 +40,6 @@ namespace Ml.Cli.WebApp.Tests
                 JsonConvert.SerializeObject(JsonConvert.DeserializeObject(compareResultContent), Formatting.Indented);
             
             fileLoader.Verify(mock => mock.WriteAllTextInFileAsync("someFolder\\\\compareFile.json", expectedResultCompareContent));
-        }
-
-        [Fact]
-        public async Task ShouldGetJsonFiles()
-        {
-            var basePath = new Mock<BasePath>("");
-            basePath.Setup(mock => mock.IsPathSecure(It.IsAny<string>())).Returns(true);
-            var fileLoader = new Mock<IFileLoader>();
-            fileLoader.Setup(mock => mock.EnumerateFiles("someFolder\\\\comparesFolder1"))
-                .Returns(new[] {"file1.json", "file2.png", "file3.json"});
-            fileLoader.Setup(mock => mock.EnumerateFiles("someFolder\\\\comparesFolder2"))
-                .Returns(new[] {"file4.png", "file5.pdf"});
-            var comparesPath = new ComparesPaths("someFolder\\\\comparesFolder1,someFolder\\\\comparesFolder2");
-            
-            var compareController = new ComparesController(fileLoader.Object, basePath.Object, comparesPath);
-
-            var result = await compareController.GetFiles() as ObjectResult;
-            
-            Assert.NotNull(result);
-            Assert.Equal(new List<string>(){"file1.json","file3.json"}, result.Value);
         }
     }
 }
